+++ rdma/test/src/bat_client_private.h
@@ -60,7 +60,7 @@ extern "C" int rdma_get_ibwin_properties(unsigned *num_ibwins,
 					 uint32_t *ibwin_size);
 extern "C" int rdma_get_msh_properties(ms_h msh,
 				       uint64_t *rio_addr,
-				       uint64_t *bytes);
+				       uint32_t *bytes);
 
 /**
  * Create an mso on the server.
